9396: Move state into parent Filters
