+++ app/javascript/components/search/Filters.js
@@ -1,3 +1,4 @@
+import isEmpty from "lodash/isEmpty";
 import React from "react";
 import PropTypes from "prop-types";
 import ButtonToolbar from "react-bootstrap/lib/ButtonToolbar";
@@ -5,6 +6,16 @@ import ButtonToolbar from "react-bootstrap/lib/ButtonToolbar";
 import FormFilter from "./FormFilter";
 
 class Filters extends React.Component {
+  constructor(props) {
+    super();
+
+    const {selectedFormIds} = props;
+    this.state = {selectedFormIds};
+
+    this.handleSubmit = this.handleSubmit.bind(this);
+    this.handleSelectForm = this.handleSelectForm.bind(this);
+  }
+
   componentDidMount() {
     // Initialize all popovers on the page.
     $(function() {
@@ -13,10 +24,29 @@ class Filters extends React.Component {
     });
   }
 
+  handleSubmit() {
+    const {selectedFormIds} = this.state;
+    const formString = isEmpty(selectedFormIds) ? "" : `form:(${selectedFormIds.join("|")})`;
+
+    // TODO: Play nicely with the other filter options instead of clobbering them.
+    window.location.href = `?search=${formString}`;
+  }
+
+  handleSelectForm(event) {
+    this.setState({selectedFormIds: [event.target.value]});
+  }
+
   render() {
+    const {allForms} = this.props;
+    const {selectedFormIds} = this.state;
+
     return (
       <ButtonToolbar className="filters">
-        <FormFilter {...this.props} />
+        <FormFilter
+          allForms={allForms}
+          onSelectForm={this.handleSelectForm}
+          onSubmit={this.handleSubmit}
+          selectedFormIds={selectedFormIds} />
       </ButtonToolbar>
     );
   }
